+++ test/util.ts
@@ -1,0 +1,3 @@
+export function as<T = any>(value: any): T {
+  return value
+}